Moving 4.x module validation to "legacy" package
