+++ web/src/main/java/com/marklogic/hub/web/service/DataHubService.java
@@ -27,8 +27,8 @@ import com.marklogic.hub.deploy.commands.LoadUserModulesCommand;
 import com.marklogic.hub.deploy.util.HubDeployStatusListener;
 import com.marklogic.hub.error.CantUpgradeException;
 import com.marklogic.hub.impl.HubConfigImpl;
+import com.marklogic.hub.legacy.validate.EntitiesValidator;
 import com.marklogic.hub.util.PerformanceLogger;
-import com.marklogic.hub.validate.EntitiesValidator;
 import com.marklogic.hub.web.auth.ConnectionAuthenticationToken;
 import com.marklogic.hub.web.exception.DataHubException;
 import com.marklogic.hub.web.listeners.DeployUserModulesListener;
